Combined ec2_availability with ec2_limits
